refactor(#3268): extract 4 subsystems from declarations.ts god-file
Verbatim moves (byte-identical emit, 39/39 IDENTICAL):
- declarations/import-collector.ts (UnifiedCollectorState subsystem)
- declarations/param-return-inference.ts (#1121 numeric inference)
- declarations/object-shape-widening.ts (empty/growable shape pre-pass)
- declarations/struct-type-registration.ts (interface/object struct reg)

Trunk 5683 -> 2417 LOC. declarations.ts re-exports moved fns for
index.ts + registry/imports.ts (zero-touch); imports back the fns the
trunk still calls internally.

Refs #3182

✓
